# crates/markitai-cli/src/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Markitai service API",
    "version": "set from the build when served",
    "summary": "Convert files and URLs to Markdown, manage saved jobs and model settings.",
    "description": "The REST API of `markitai serve`. Loopback clients are trusted; other clients send the startup token as `Authorization: Bearer` (preferred) or `?token=`. Every error body is `{detail, code, reason}`; present text by `reason`, then `code`. See docs/serve.md, docs/service-settings.md and docs/provider-management.md for the contracts this document summarizes.",
    "license": { "name": "MIT", "identifier": "MIT" }
  },
  "servers": [{ "url": "/", "description": "The service that serves this document" }],
  "security": [{}, { "bearerAuth": [] }, { "tokenQuery": [] }],
...
  "paths": {
...
    "/api/settings/llm": {
      "get": {
        "tags": ["settings"],
        "operationId": "getSettings",
        "summary": "Revision and secret-free saved and detected deployments",
        "responses": {
          "200": { "description": "Settings", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SettingsView" } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/settings/llm/providers": {
      "get": {
        "tags": ["settings"],
        "operationId": "listProviders",
        "summary": "Environment, common and saved provider cards",
        "parameters": [{ "name": "refresh", "in": "query", "required": false, "description": "Accepted for compatibility; the list is computed on every request", "schema": { "type": "boolean" } }],
        "responses": {
          "200": { "description": "Provider cards", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProviderList" } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/settings/llm/providers/{id}/credentials": {
      "get": {
        "tags": ["settings"],
        "operationId": "getProviderCredentials",
        "summary": "The stored, editable fields of one saved connection",
        "parameters": [{ "$ref": "#/components/parameters/SettingsId" }],
        "responses": {
          "200": { "description": "Stored fields (an `env:NAME` reference stays unresolved)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProviderCredentials" } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/settings/llm/models": {
      "post": {
        "tags": ["settings"],
        "operationId": "addModel",
        "summary": "Add one deployment (compatible with the legacy single-model form)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewDeployment" } } } },
        "responses": { "200": { "$ref": "#/components/responses/Settings" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "422": { "$ref": "#/components/responses/Error" }, "500": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/settings/llm/models/{id}": {
      "put": {
        "tags": ["settings"],
        "operationId": "updateLegacyModel",
        "summary": "Change the one deployment of an unambiguous routing group",
        "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The routing group (`model_name`)", "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeploymentPatch" } } } },
        "responses": { "200": { "$ref": "#/components/responses/Settings" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "422": { "$ref": "#/components/responses/Error" }, "500": { "$ref": "#/components/responses/Error" } }
      },
      "delete": {
        "tags": ["settings"],
        "operationId": "deleteLegacyModel",
        "summary": "Delete the one deployment of an unambiguous routing group",
        "parameters": [{ "name": "id", "in": "path", "required": true, "description": "The routing group (`model_name`)", "schema": { "type": "string" } }],
        "responses": { "200": { "$ref": "#/components/responses/Settings" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "500": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/settings/llm/deployments/batch": {
      "post": {
        "tags": ["settings"],
        "operationId": "addDeployments",
        "summary": "Atomically add 1–50 deployments",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeploymentBatch" } } } },
        "responses": { "200": { "$ref": "#/components/responses/Settings" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "422": { "$ref": "#/components/responses/Error" }, "500": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/settings/llm/deployments/{id}": {
      "patch": {
        "tags": ["settings"],
        "operationId": "updateDeployment",
        "summary": "Change one deployment by its stable identity",
        "parameters": [{ "$ref": "#/components/parameters/SettingsId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/DeploymentPatch" }], "required": ["expected_revision"] } } } },
        "responses": { "200": { "$ref": "#/components/responses/Settings" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "422": { "$ref": "#/components/responses/Error" }, "500": { "$ref": "#/components/responses/Error" } }
      },
      "delete": {
        "tags": ["settings"],
        "operationId": "deleteDeployment",
        "summary": "Delete one deployment; its connection stays saved",
        "parameters": [{ "$ref": "#/components/parameters/SettingsId" }, { "$ref": "#/components/parameters/ExpectedRevision" }],
        "responses": { "200": { "$ref": "#/components/responses/Settings" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "422": { "$ref": "#/components/responses/Error" }, "500": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/settings/llm/providers/{id}": {
      "patch": {
        "tags": ["settings"],
        "operationId": "updateProvider",
        "summary": "Change a saved connection and every deployment linked to it",
        "parameters": [{ "$ref": "#/components/parameters/SettingsId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProviderPatch" } } } },
        "responses": { "200": { "$ref": "#/components/responses/Settings" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "422": { "$ref": "#/components/responses/Error" }, "500": { "$ref": "#/components/responses/Error" } }
      },
      "delete": {
        "tags": ["settings"],
        "operationId": "deleteProvider",
        "summary": "Delete a saved connection and its deployments",
        "parameters": [{ "$ref": "#/components/parameters/SettingsId" }, { "$ref": "#/components/parameters/ExpectedRevision" }],
        "responses": { "200": { "$ref": "#/components/responses/Settings" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "422": { "$ref": "#/components/responses/Error" }, "500": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/settings/llm/config/open": {
      "post": {
        "tags": ["settings"],
        "operationId": "openConfig",
        "summary": "Ask the server host's system opener to open the configuration file",
        "responses": {
          "204": { "description": "Opened" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "The startup token (printed at start, or MARKITAI_SERVE_TOKEN). Loopback peers need none." },
      "tokenQuery": { "type": "apiKey", "in": "query", "name": "token", "description": "The startup token as a query parameter; accepted for compatibility, but it can end up in logs and history, so prefer the header." },
      "ticketQuery": { "type": "apiKey", "in": "query", "name": "ticket", "description": "A single-use download ticket from POST /api/download-tickets, valid for one GET of the path it was issued for within 60 seconds." }
    },
    "parameters": {
...
      "SettingsId": { "name": "id", "in": "path", "required": true, "description": "A deployment or provider identity from the settings views", "schema": { "type": "string" } },
      "ExpectedRevision": { "name": "expected_revision", "in": "query", "required": true, "description": "The revision the change was drafted against", "schema": { "type": "string" } }
    },
    "responses": {
      "Error": { "description": "Refusal", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Settings": { "description": "The settings after the change", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SettingsView" } } } }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["detail", "code", "reason"],
        "properties": {
          "detail": { "description": "The service's English message, or an object for settings conflicts that carry `detail.code`", "oneOf": [{ "type": "string" }, { "type": "object" }] },
          "code": { "type": "string", "enum": ["bad_request", "unauthorized", "forbidden", "not_found", "method_not_allowed", "conflict", "payload_too_large", "invalid_request", "rate_limited", "unavailable", "server_error"] },
          "reason": {
            "type": "string",
            "description": "Stable machine code of the specific cause. The list is the set the service writes; new codes may be added.",
            "examples": ["ambiguous_legacy_model_name", "archive_too_large", "batch_pending", "config_changed", "config_missing", "config_too_large", "discovery_timeout", "empty_job", "enhancement_failed", "file_not_found", "file_too_large", "form_field_too_large", "history_empty", "host_not_allowed", "internal_error", "invalid_allowed_host", "invalid_json", "invalid_multipart", "invalid_options", "invalid_provider_request", "invalid_retry_body", "invalid_settings_request", "invalid_ticket_path", "invalid_url", "invalid_urls", "item_busy", "item_not_found", "job_id_collision", "job_not_found", "job_not_running", "job_running", "json_required", "llm_unavailable", "method_not_allowed", "no_output", "no_source", "not_retryable", "nothing_to_stop", "origin_not_allowed", "output_conflict", "output_identity_conflict", "persistence_failed", "provider_busy", "provider_unavailable", "remote_url_forbidden", "request_too_large", "result_too_large", "result_unavailable", "route_not_found", "settings_durability_unknown", "settings_entry_missing", "settings_forbidden", "settings_io_failed", "settings_read_only", "shutting_down", "stale_revision", "ticket_invalid", "token_required", "too_many_items", "too_many_tickets", "unknown_preset", "unsupported_url_scheme", "upload_missing"]
          }
        }
      },
...
      "Deployment": {
        "type": "object",
        "required": ["deployment_id", "routing_group", "model", "weight", "api_key_configured", "api_base_configured", "api_base", "persisted"],
        "properties": {
          "deployment_id": { "type": "string" },
          "routing_group": { "type": "string" },
          "model": { "type": "string" },
          "weight": { "type": "integer" },
          "api_key_configured": { "type": "boolean" },
          "api_base_configured": { "type": "boolean" },
          "api_base": { "type": ["string", "null"], "description": "The endpoint's HTTP(S) origin only" },
          "persisted": { "type": "boolean" }
        }
      },
      "SettingsView": {
        "type": "object",
        "required": ["configured", "routable", "source", "config_path", "config_origin", "revision", "deployments", "detected"],
        "properties": {
          "configured": { "type": "boolean" },
          "routable": { "type": "boolean" },
          "source": { "type": "string", "enum": ["config", "detected", "none"] },
          "config_path": { "type": "string" },
          "config_origin": { "type": "string" },
          "revision": { "type": "string" },
          "deployments": { "type": "array", "items": { "$ref": "#/components/schemas/Deployment" } },
          "detected": { "type": "array", "items": { "$ref": "#/components/schemas/Deployment" } }
        }
      },
      "ProviderCard": {
        "type": "object",
        "required": ["id", "provider", "label", "kind", "status", "source", "supports_discovery"],
        "properties": {
          "id": { "type": "string" },
          "provider_id": { "type": "string" },
          "deployment_id": { "type": "string" },
          "provider": { "type": "string", "description": "A model prefix (`openai`, `groq`, …) or `custom` for an OpenAI-compatible endpoint" },
          "label": { "type": "string" },
          "kind": { "type": "string", "enum": ["environment", "configured", "common"] },
          "status": { "type": "string", "examples": ["ready", "needs_credentials", "unknown", "disabled"] },
          "source": { "type": "string", "description": "The environment variable, `config` or `built_in`" },
          "credential": { "type": "string" },
          "api_key_configured": { "type": "boolean" },
          "api_base_configured": { "type": "boolean" },
          "api_base": { "type": ["string", "null"] },
          "default_base": { "type": ["string", "null"], "description": "The provider's documented endpoint; null when every server has its own" },
          "key_variable": { "type": ["string", "null"], "description": "The environment variable read for the key" },
          "key_optional": { "type": "boolean", "description": "A local server that works without a key" },
          "model_count": { "type": "integer" },
          "supports_discovery": { "type": "boolean", "description": "False when the provider has no model list endpoint: enter model IDs by hand" }
        }
      },
      "ProviderList": {
        "type": "object",
        "required": ["providers"],
        "properties": { "providers": { "type": "array", "items": { "$ref": "#/components/schemas/ProviderCard" } } }
      },
      "ProviderCredentials": {
        "type": "object",
        "required": ["api_key", "api_base", "api_base_placeholder"],
        "properties": {
          "api_key": { "type": ["string", "null"] },
          "api_base": { "type": ["string", "null"] },
          "api_base_placeholder": { "type": ["string", "null"] }
        }
      },
      "NewDeployment": {
        "type": "object",
        "required": ["model_name", "model"],
        "additionalProperties": false,
        "properties": {
          "model_name": { "type": "string", "description": "Routing group" },
          "model": { "type": "string", "examples": ["openai/gpt-5.6-luna"] },
          "provider": { "type": "string" },
          "api_key": { "type": "string", "description": "A literal key or `env:NAME`" },
          "api_base": { "type": "string" },
          "weight": { "type": "integer", "minimum": 0, "default": 1 },
          "credential_provider_id": { "type": "string" },
          "credential_deployment_id": { "type": "string" },
          "expected_revision": { "type": "string" }
        }
      },
      "DeploymentBatch": {
        "type": "object",
        "required": ["expected_revision", "deployments"],
        "additionalProperties": false,
        "properties": {
          "expected_revision": { "type": "string" },
          "deployments": { "type": "array", "minItems": 1, "maxItems": 50, "items": { "$ref": "#/components/schemas/NewDeployment" } }
        }
      },
      "DeploymentPatch": {
        "type": "object",
        "additionalProperties": false,
        "description": "Omitted fields are kept; null clears api_key or api_base.",
        "properties": {
          "model_name": { "type": "string" },
          "model": { "type": "string" },
          "api_key": { "type": ["string", "null"] },
          "api_base": { "type": ["string", "null"] },
          "weight": { "type": "integer", "minimum": 0 },
          "expected_revision": { "type": "string" }
        }
      },
      "ProviderPatch": {
        "type": "object",
        "required": ["expected_revision"],
        "additionalProperties": false,
        "properties": {
          "api_key": { "type": ["string", "null"] },
          "api_base": { "type": ["string", "null"] },
          "expected_revision": { "type": "string" }
        }
      },
...
    }
  }
}